pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ftpc"
version = "0.1.0"
description = "TUI file transfer client with support for multiple storage backends"
authors = [
    {name = "Edoardo Annunziata", email = "annunziata.edoardo@outlook.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console :: Curses",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Topic :: Utilities",
]
dependencies = [
    "windows-curses; sys_platform == 'win32'",
]

[project.urls]
Homepage = "https://github.com/edoannunziata/ftpc"
Repository = "https://github.com/edoannunziata/ftpc"

[project.optional-dependencies]
azure = [
    "azure-storage-file-datalake>=12.8.0",
    "azure-identity>=1.10.0",
]
blob = [
    "azure-storage-blob>=12.14.0",
    "azure-identity>=1.10.0",
]
s3 = [
    "boto3>=1.28.0",
]
sftp = [
    "paramiko>=3.0.0",
]
proxy = [
    "PySocks>=1.7.1",
]
async = [
    "aiofiles>=23.2.1",
    "aioftp>=0.21.4",
    "asyncssh>=2.14.2",
    "aioboto3>=12.3.0",
]
all = [
    "azure-storage-file-datalake>=12.8.0",
    "azure-storage-blob>=12.14.0",
    "azure-identity>=1.10.0",
    "boto3>=1.28.0",
    "paramiko>=3.0.0",
    "PySocks>=1.7.1",
    "aiofiles>=23.2.1",
    "aioftp>=0.21.4",
    "asyncssh>=2.14.2",
    "aioboto3>=12.3.0",
]
test = [
    "coverage>=7.0.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
ftpc = "ftpc.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ftpc*"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
    "boto3.*",
    "botocore.*",
    "paramiko.*",
    "socks",
    "azure.*",
    "aioftp.*",
    "asyncssh.*",
    "aioboto3.*",
    "aiofiles.*",
]
ignore_missing_imports = true
